Move genotype file reading and parsing logic to `Reader`
